// src/tenable/types.ts
export interface RecentScanSummary {
  id: number;
  legacy: boolean;
  permissions: number;
  type: string;
  read: boolean;
  last_modification_date: number;
  creation_date: number;
  status: ScanStatus;
  uuid: string;
  shared: boolean;
  user_permissions: number;
  owner: string;
  schedule_uuid: string;
  timezone?: string;
  rrules?: string;
  starttime?: string;
  enabled: boolean;
  control: boolean;
  name: string;
}

export enum ScanStatus {
  Completed = 'completed',
  Aborted = 'aborted',
  Empty = 'empty',
  Imported = 'imported',
  Pending = 'pending',
  Running = 'running',
  Resuming = 'resuming',
  Canceling = 'canceling',
  Canceled = 'canceled',
  Pausing = 'pausing',
  Paused = 'paused',
  Stopping = 'stopping',
  Stopped = 'stopped',
}

// -- https://cloud.tenable.com/scans/:scanId
//    https://developer.tenable.com/reference#scans-details

/**
 * Details of a scan, providing a lot more information about the current/recent
 * execution.
 */
export interface RecentScanDetail {
  /**
   * Details about the scan itself.
   */
  info: ScanInfo;

  /**
   * The hosts that were included in the current/recent exection of the scan,
   * `undefined` when there are no hosts returned by the API.
   */
  hosts?: ScanHost[];

  /**
   * Vulnerabilities found during the current/recent execution of the scan,
   * `undefined` when there are no vulnerabilities returned by the API.
   */
  vulnerabilities?: ScanVulnerabilitySummary[];
}

export interface ScanInfo {
  owner: string;
  name: string;
  no_target: boolean;
  folder_id: number;
  control: boolean;
  user_permissions: number;
  schedule_uuid: string;
  edit_allowed: boolean;
  scanner_name: string;
  policy: string;
  shared?: string;
  object_id: number;
  hostcount: number;
  uuid: string;
  status: string;
  scan_type: string;
  targets: string;
  alt_targets_used: boolean;
  'pci-can-upload': boolean;
  is_archived: boolean;
  timestamp: number;
  haskb: boolean;
  hasaudittrail: boolean;
}

export interface ScanHost {
  uuid: string;
  critical: number;
  high: number;
  medium: number;
  low: number;
  info: number;
  severity: number;
  host_id: number;
  hostname: string;
}

export interface ScanVulnerabilitySummary {
  count: number;
  plugin_family: string;
  plugin_id: number;
  plugin_name: string;
  severity: number;
}
// ...
export interface ScansResponse {
  folders: Folder[];
  scans: RecentScanSummary[];
  timestamp: number;
}

interface Folder {
  unread_count: number;
  custom: number;
  default_tag: number;
  type: string;
  name: string;
  id: number;
}

export interface ScanResponse {
  info: ScanInfo;
  hosts?: ScanHost[];
  vulnerabilities: ScanVulnerabilitySummary[];
}